UX Stuff

+++ packages/client/src/components/reset.tsx
@@ -7,8 +7,10 @@ const ResetButton = () => {
   })
 
   const handleReset = () => {
-    localStorage.clear()
-    sendJsonMessage({ type: "end" } as Message)
+    if (confirm("Are you sure you want to reset the game?")) {
+      localStorage.clear()
+      sendJsonMessage({ type: "end" } as Message)
+    }
   }
 
   return (
@@ -16,7 +18,7 @@ const ResetButton = () => {
       onClick={handleReset}
       className="bg-red-500 hover:bg-red-700 text-white fixed m-2 font-bold py-2 px-2 rounded"
     >
-      Reset Game
+      End Game
     </button>
   )
 }
